Refactors Contansts in a Package
