+++ benchflow-test-manager/application/src/test/java/cloud/benchflow/testmanager/helpers/constants/TestConstants.java
@@ -1,4 +1,4 @@
-package cloud.benchflow.testmanager.helpers;
+package cloud.benchflow.testmanager.helpers.constants;
 
 import static cloud.benchflow.testmanager.constants.BenchFlowConstants.MODEL_ID_DELIMITER;
